# microservices/api-index-api/src/main/resources/openapi/v1-api-index-api.yml
openapi: 3.1.2
info:
  title: API Index-API
  description: A microservice for managing indexed API's
  version: 1.0.0
  x-api-name: api-index-api
  x-service-name: api-index-api
paths:
  /api/rest/v1/apis:
    post:
      operationId: ingestApi
      summary: Ingest an API specification
      description: |
        Uploads and indexes an API specification for Snow-White.
        If a stable API with the same service name, API name, and version already exists, the request is rejected with HTTP 409 and the existing specification remains unchanged.
        Prerelease specifications may be re-uploaded to override an existing prerelease entry with the same identity.
      tags:
        - api-index
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/ApiInformation.yml'
      responses:
        '201':
          description: The API specification was successfully ingested and indexed.
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: './components/Error.yml'
        '409':
          description: An API specification with the same service name, API name, and version already exists.
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: './components/Error.yml'
    get:
      operationId: getAllApis
      summary: List ingested APIs
      description: Returns an overview of all API specifications currently indexed by the system.
      tags:
        - api-index
      parameters:
        - name: page
          in: query
          description: Page number (0-based index).
          required: false
          schema:
            type: integer
            minimum: 0
        - name: size
          in: query
          description: Number of items per page.
          required: false
          schema:
            type: integer
            minimum: 1
        - name: sort
          in: query
          description: |
            Sorting criteria in the format: `property,(asc|desc)`.
            One of `apiName`, `apiType`, `apiVersion`, `serviceName`; any other property or a
            malformed value is rejected with `400`.
            Defaults to `serviceName,asc`. Every order is completed by `serviceName`, `apiName`,
            `apiVersion` — the key identifying an API — so paging is reproducible.
            `apiVersion` orders as a string, so `1.10.0` precedes `1.9.0`.
          required: false
          schema:
            type: string
            example: serviceName,asc
        - name: serviceName
          in: query
          required: false
          schema:
            type: string
            maxLength: 64
          description: 'Filter results to APIs belonging to this service name (combine with apiName for cascading)'
        - name: apiName
          in: query
          required: false
          schema:
            type: string
            maxLength: 64
          description: 'Filter results to APIs with this API name (combine with serviceName for cascading)'
      responses:
        '200':
          description: The list of ingested API specifications was successfully retrieved.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: './components/ApiInformation.yml'
        '400':
          description: 'The sort parameter names a property this listing cannot order by, or does not parse'
          content:
            application/json:
              schema:
                $ref: './components/Error.yml'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: './components/Error.yml'
  /api/rest/v1/apis/meta/service-names:
    get:
      operationId: getAllServiceNames
      summary: List indexed service names
      description: Returns a deduplicated list of all service names currently present in the index.
      tags:
        - api-index
      responses:
        '200':
          description: The list of indexed service names was successfully retrieved.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: './components/Error.yml'
  /api/rest/v1/apis/meta/api-names:
    get:
      operationId: getAllApiNames
      summary: List indexed API names
      description: Returns a deduplicated list of all API names currently present in the index, optionally filtered to a specific service.
      tags:
        - api-index
      parameters:
        - name: serviceName
          in: query
          required: false
          schema:
            type: string
            maxLength: 64
          description: 'Filter results to API names belonging to this service name'
      responses:
        '200':
          description: The list of indexed API names was successfully retrieved.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: './components/Error.yml'
  /api/rest/v1/apis/{otelServiceName}/{apiName}/{apiVersion}:
    get:
      operationId: getApiDetails
      summary: Get API details
      description: Retrieves detailed information for a specific ingested API specification identified by service name, API name, and version.
      tags:
        - api-index
      parameters:
        - name: otelServiceName
          in: path
          required: true
          schema:
            type: string
        - name: apiName
          in: path
          required: true
          schema:
            type: string
        - name: apiVersion
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The requested API specification was found and returned.
          content:
            application/json:
              schema:
                $ref: './components/ApiInformation.yml'
        '404':
          description: No API specification exists for the given service name, API name, and version.
          content:
            application/json:
              schema:
                $ref: './components/Error.yml'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: './components/Error.yml'
  /api/rest/v1/apis/{otelServiceName}/{apiName}/{apiVersion}/raw:
    get:
      operationId: getRawApiContent
      summary: Get raw prerelease API specification
      description: Returns the raw specification content for a prerelease API. Returns 404 if not found or not a prerelease.
      tags:
        - api-index
      parameters:
        - name: otelServiceName
          in: path
          required: true
          schema:
            type: string
        - name: apiName
          in: path
          required: true
          schema:
            type: string
        - name: apiVersion
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The raw prerelease API specification content.
          content:
            application/yaml:
              schema:
                type: string
            text/plain:
              schema:
                type: string
        '404':
          description: The API specification does not exist or is not a prerelease.
          content:
            application/json:
              schema:
                $ref: './components/Error.yml'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: './components/Error.yml'
  /api/rest/v1/apis/{otelServiceName}/{apiName}/{apiVersion}/exists:
    get:
      operationId: checkApiExists
      summary: Check API existence
      description: Lightweight endpoint to quickly verify whether a specific API specification has already been ingested.
      tags:
        - api-index
      parameters:
        - name: otelServiceName
          in: path
          required: true
          schema:
            type: string
        - name: apiName
          in: path
          required: true
          schema:
            type: string
        - name: apiVersion
          in: path
          required: true
          schema:
            type: string
        - name: includePrereleases
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: The API specification exists.
        '404':
          description: The API specification does not exist.
          content:
            application/json:
              schema:
                $ref: './components/Error.yml'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: './components/Error.yml'